Hack ``version_to_segments`` to allow invalid type
NOT FOUND integration test makes up a type that the server can't
possibly know about.  Of course, the client doesn't know about it either
so can't formulate a request.  Hack things up a bit in that test to
allow the server to be encountered on the server rather than the client.